[FEAT] Imply Recv and Send Buffer

+++ DummyClient/pch.h
@@ -14,4 +14,4 @@
 
 #include "../TetrisServer/pch.h"
 
-using PacketVariant = variant<uint8_t, uint16_t, double, string, uint16_t*, double*>;
+using PacketVariant = variant<uint8_t, uint16_t, uint64_t, string, uint16_t*, uint64_t*>;
